+++ internal/search/subrepoperms/sub_repo_perms_job_test.go
@@ -1,4 +1,4 @@
-package run
+package subrepoperms
 
 import (
 	"context"
